maze/requirements.txt
numpy
imageio
pillow
streamlit-image-select